requirements.txt
streamlit_lottie==0.0.2
pygsheets==2.0.5
plotly==5.1.0
streamlit_aggrid==0.2.1
requests==2.25.1
pandas==1.2.4
streamlit==0.88.0
Pillow==8.3.2
